Add files via upload
Here i webscrap one amazon product data and take the product review to check the product rating
justified or not. Here We take Both low star rating reviews and high star rating views. Using wordcloud we
check the most used word in the reviews. Using Natural Language Processing (NLP) we predict that actually the product is good or not